refactor: moved parameters into its own module
