+++ cognite_toolkit/_cdf_tk/load/_cls_parameters/get_type_hints.py
@@ -7,6 +7,9 @@ import typing
 from collections import Counter
 from typing import Any, get_type_hints
 
+from cognite.client.data_classes import TransformationScheduleWrite
+from cognite.client.data_classes.capabilities import UnknownAcl
+
 
 class _TypeHints:
     """
@@ -15,12 +18,15 @@ class _TypeHints:
     After Python 3.10, type hints are treated as strings, so we need to evaluate them to get the actual type.
     """
 
+    _error_msg: typing.ClassVar[str] = "Please extend this function to support generating fake data for this type"
+
     @classmethod
     def get_concrete_classes(cls, resource_cls: type) -> list[type]:
         """If the resource class is a ABC class, then, this function will return all the concrete classes
         that are subclasses of the resource class."""
         is_base_class = inspect.isclass(resource_cls) and any(base is abc.ABC for base in resource_cls.__bases__)
-        if not is_base_class:
+        if not is_base_class or resource_cls is TransformationScheduleWrite:
+            # TransformationScheduleWrite is wrongly marked as ABC, but it is a concrete class
             # Easy case
             return [resource_cls]
         concrete_classes = []
@@ -29,7 +35,11 @@ class _TypeHints:
             cls_ = to_check.pop()
             to_check.extend(cls_.__subclasses__())
             is_base_class = inspect.isclass(cls_) and any(base is abc.ABC for base in cls_.__bases__)
-            if not is_base_class:
+            # UnknownAcl is a special case, it is concrete class, but cannot be instantiated easily
+            if not is_base_class and cls_ is not UnknownAcl:
+                concrete_classes.append(cls_)
+            if cls_ is TransformationScheduleWrite:
+                # TransformationScheduleWrite is wrongly marked as ABC, but it is a concrete class
                 concrete_classes.append(cls_)
         return concrete_classes
 
@@ -48,7 +58,7 @@ class _TypeHints:
             return {}
         try:
             type_hint_by_name = get_type_hints(resource_cls.__init__, localns=cls._type_checking())  # type: ignore[misc]
-        except TypeError:
+        except (TypeError, NameError):
             # Python 3.10 Type hints cannot be evaluated with get_type_hints,
             # ref https://stackoverflow.com/questions/66006087/how-to-use-typing-get-type-hints-with-pep585-in-python3-8
             resource_module_vars = vars(importlib.import_module(resource_cls.__module__))
@@ -68,6 +78,7 @@ class _TypeHints:
         import numpy as np
         import numpy.typing as npt
         from cognite.client import CogniteClient
+        from cognite.client.data_classes.data_modeling import ContainerId, ViewId
 
         return {
             "CogniteClient": CogniteClient,
@@ -76,6 +87,10 @@ class _TypeHints:
             "NumpyInt64Array": npt.NDArray[np.int64],
             "NumpyFloat64Array": npt.NDArray[np.float64],
             "NumpyObjArray": npt.NDArray[np.object_],
+            # These are imported locally in the HasData filter classes
+            # and are thus not in the module or class namespace
+            "ViewId": ViewId,
+            "ContainerId": ContainerId,
         }
 
     @classmethod
@@ -97,7 +112,7 @@ class _TypeHints:
         annotation = annotation.replace("SequenceNotStr", "Sequence")
         try:
             return eval(annotation, resource_module_vars, local_vars)
-        except TypeError:
+        except (TypeError, NameError):
             # Python 3.10 Type Hint
             return cls._type_hint_3_10_to_8(annotation, resource_module_vars, local_vars)
