+++ cognite_toolkit/_cdf_tk/validation/read_yaml.py
@@ -7,11 +7,8 @@ import re
 import sys
 import types
 import typing
-from collections.abc import Hashable, Iterable, MutableSet
-from dataclasses import dataclass
-from functools import total_ordering
 from pathlib import Path
-from typing import Any, ClassVar, Generic, TypeVar, get_origin
+from typing import Any, get_origin
 
 from cognite.client.data_classes._base import CogniteObject
 from cognite.client.utils._text import to_camel_case, to_snake_case
@@ -179,187 +176,4 @@ def validate_data_set_is_set(
     return warning_list
 
 
-# These are internal classes that are used by the validate parameter functions.
-@total_ordering
-@dataclass(frozen=True)
-class Parameter:
-    path: tuple[str | int, ...]
-    type: str
-
-    def __lt__(self, other: Parameter) -> bool:
-        if not isinstance(other, Parameter):
-            return NotImplemented
-        return self.path < other.path
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, Parameter):
-            return NotImplemented
-        return self.path == other.path and self.type == other.type
-
-
-@dataclass(frozen=True)
-class ParameterSpec(Parameter):
-    is_required: bool
-    _is_nullable: bool | None = None
-
-    @property
-    def is_nullable(self) -> bool:
-        return self._is_nullable or not self.is_required
-
-
-@dataclass(frozen=True)
-class ParameterValue(Parameter):
-    value: str | int | float | bool | list | dict
-
-
-T_Parameter = TypeVar("T_Parameter", bound=Parameter)
-
-
-class ParameterSet(Hashable, MutableSet, Generic[T_Parameter]):
-    def __init__(self, iterable: Iterable[T_Parameter] = ()) -> None:
-        self.data: set[T_Parameter] = set(iterable)
-
-    def __hash__(self) -> int:
-        return hash(self.data)
-
-    def __contains__(self, value: object) -> bool:
-        return value in self.data
-
-    def __iter__(self) -> typing.Iterator[T_Parameter]:
-        return iter(self.data)
-
-    def __len__(self) -> int:
-        return len(self.data)
-
-    def __repr__(self) -> str:
-        return repr(self.data)
-
-    def add(self, item: T_Parameter) -> None:
-        self.data.add(item)
-
-    def discard(self, item: T_Parameter) -> None:
-        self.data.discard(item)
-
-    def update(self, other: ParameterSet[T_Parameter]) -> None:
-        self.data.update(other.data)
-
-
-class ParameterSpecSet(ParameterSet[ParameterSpec]):
-    def __init__(self, iterable: Iterable[ParameterSpec] = ()) -> None:
-        super().__init__(iterable)
-        self.is_complete = True
-
-    @property
-    def required(self) -> ParameterSet[ParameterSpec]:
-        return ParameterSet[ParameterSpec](parameter for parameter in self if parameter.is_required)
-
-    def update(self, other: ParameterSet[ParameterSpec]) -> None:
-        if isinstance(other, ParameterSpecSet):
-            self.is_complete &= other.is_complete
-        super().update(other)
-
-
-class _TypeHint:
-    _BASE_TYPES: ClassVar[set[str]] = {t.__name__ for t in (str, int, float, bool)}
-    _CONTAINER_TYPES: ClassVar[set[str]] = {t.__name__ for t in (list, dict)}
-
-    def __init__(self, raw: Any) -> None:
-        self.raw = raw
-        self._container_type = get_origin(raw)
-        self.args = typing.get_args(raw)
-        self._is_nullable = False
-        if self._container_type in [types.UnionType, typing.Union] and self.args:
-            inner_container = typing.get_origin(self.args[0])
-            if inner_container:
-                self._is_nullable = any(arg in [None, types.NoneType] for arg in self.args)
-                self._container_type = inner_container
-                self.args = typing.get_args(self.args[0])
-
-    def __str__(self) -> str:
-        if self._container_type and self._container_type not in [types.UnionType, typing.Union]:
-            value = self._container_type.__name__
-        else:
-            value = self.arg.__name__
-        if value in self._CONTAINER_TYPES or value in self._BASE_TYPES:
-            return value
-        elif value == "Literal":
-            return "str"
-        return "dict"
-
-    @property
-    def arg(self) -> Any:
-        if (self._container_type in [typing.Union, types.UnionType]) and self.args:
-            return self.args[0]
-        return self.raw
-
-    @property
-    def is_base_type(self) -> bool:
-        return str(self) in self._BASE_TYPES
-
-    @property
-    def is_nullable(self) -> bool:
-        return self._is_nullable or any(arg is None or arg is types.NoneType for arg in self.args)
-
-    @property
-    def is_class(self) -> bool:
-        return inspect.isclass(self.arg)
-
-    @property
-    def is_dict_type(self) -> bool:
-        return self._container_type is dict
-
-    @property
-    def is_list_type(self) -> bool:
-        return self._container_type is list
-
-    def __repr__(self) -> str:
-        return repr(self.arg)
-
-
 # Todo: Add caching at this level
-def read_parameter_from_init_type_hints(cls_: type) -> ParameterSpecSet:
-    return _read_parameter_from_init_type_hints(cls_, tuple(), set())
-
-
-def _read_parameter_from_init_type_hints(cls_: type, path: tuple[str | int, ...], seen: set[str]) -> ParameterSpecSet:
-    path = tuple() if path is None else path
-    parameter_set = ParameterSpecSet()
-    if not hasattr(cls_, "__init__"):
-        return parameter_set  # type: ignore[misc]
-
-    classes = _TypeHints.get_concrete_classes(cls_)
-    seen.update(cls_.__name__ for cls_ in classes)
-    type_hints_by_name = _TypeHints.get_type_hints_by_name(classes)
-    parameters = {k: v for cls in classes for k, v in inspect.signature(cls.__init__).parameters.items()}  # type: ignore[misc]
-    for name, parameter in parameters.items():
-        if name == "self" or parameter.kind is parameter.VAR_POSITIONAL or parameter.kind is parameter.VAR_KEYWORD:
-            continue
-        try:
-            hint = _TypeHint(type_hints_by_name[name])
-        except KeyError:
-            # Missing type hint
-            parameter_set.is_complete = False
-            continue
-        is_required = parameter.default is inspect.Parameter.empty
-        is_nullable = hint.is_nullable
-        parameter_set.add(ParameterSpec((*path, name), str(hint), is_required, is_nullable))
-        if hint.is_base_type or str(hint) in seen:
-            continue
-        elif hint.is_dict_type:
-            key, value = hint.args
-            dict_set = _read_parameter_from_init_type_hints(value, (*path, name), seen)
-            parameter_set.update(dict_set)
-        if hint.is_list_type:
-            item_hint = _TypeHint(hint.args[0])
-            if item_hint.is_base_type:
-                parameter_set.add(ParameterSpec((*path, name, 0), str(item_hint), is_required, is_nullable))
-            elif hint.args[0].__name__ in seen:
-                parameter_set.add(ParameterSpec((*path, name, 0), "dict", is_required, is_nullable))
-            else:
-                list_set = _read_parameter_from_init_type_hints(hint.args[0], (*path, name, 0), seen)
-                parameter_set.update(list_set)
-        elif hint.is_class:
-            cls_set = _read_parameter_from_init_type_hints(hint.arg, (*path, name), seen)
-            parameter_set.update(cls_set)
-
-    return parameter_set
